refactor: move out plan into its own file
